feat: use common formatRange function
