+++ libs/utils/src/lib/format/index.ts
@@ -1,4 +1,5 @@
 export * from './date';
 export * from './number';
+export * from './range';
 export * from './size';
 export * from './strings';
